requirements.txt
Flask==1.1.2
Flask-Bcrypt==0.7.1
Flask-Cors==3.0.9
Flask-Migrate==2.5.3
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.4
numpy==1.19.3
pandas==1.1.4
python-dotenv==0.15.0
requests==2.25.1
scikit-learn==0.24.2
scipy==1.5.4
tqdm==4.54.1
urllib3==1.26.4
PyJWT==2.8.0
# Compatibility fixes for Flask 1.1.2
Jinja2<3.1
MarkupSafe<2.1
itsdangerous<2.1
Werkzeug<2.0
SQLAlchemy<1.4
